JRR[MINOR: Styled action buttons in each form.
All the forms were unified with the action button to the right (more
natural place to be selected with the pointer), and highlighted with the
combining blue color and shape of the button.
It makes it visually clear what will happen by pressing the ENTER key
while in the form.